Challenge 12 Work done
I did a whoooole ton of stuff with this one at least for the pathfinding and whatnot. So the basis of the challenge wanted me to go and make the enemy go between 2 points on a loop.
I thought that was too simple so I made it go between the points with full animation for running and pausing, speaking of I made it pause when it reaches the points for a specified amount of time. I also made it target the player as well when the player is in a certain range (loses aggression when player is out of range.) I also made the sprite flip properly when going one way or another.

I think if any more changes are to be made on this section I am going to want to tweak the setting on the run to pause movement (might just need to make it not have exit time on the animation) and a few other spots I cannot think of that I noticed might help.

Really long comments under this change but I actually felt joy out of doing this one so that's a nice change of pace.