Large amount of work done.
Did a ton of level work, script work, character work, and made transitions between scenes work. Also created a full menu with music settings, a exit and resume by pressing escape.
I am missing a couple items here but this should be it for the most part, just wrapping up the final work to submit it for the portfolio.